==> 5_1_dressing_percentages_protein.Rmd <==
---
title: "Dressing percentages etc"
author: "JP"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



# Get dressing percentages from GLEAM v3
https://view.officeapps.live.com/op/view.aspx?src=https%3A%2F%2Fwww.fao.org%2Ffileadmin%2Fuser_upload%2Fgleam%2Fdocs%2FGLEAM_3.0_Supplement_S1.xlsx&wdOrigin=BROWSELINK

Regions: NA (North America), RUS (Russian Federation), WE (Western Europe), EE (Eastern Europe), NENA (Near East and North Africa), ESEA (East and Southeast Asia), OCE (Oceania), SA (South Asia), LAC (Latin America and the Caribbean) and SSA (Sub-Saharan Africa)

```{r}

# Creating a tibble with the provided data
dressing_percentages <- tibble(
  Category = c("Dairy cattle", "Dairy cattle", "Dairy cattle", 
               "Beef cattle", "Beef cattle", "Beef cattle", 
               "Buffaloes", "Buffaloes", "Buffaloes"),
  Type = c("Adult and replacement female", "Adult and replacement male", "Surplus female and male", 
           "Adult and replacement female", "Adult and replacement male", "Surplus female and male",
           "Adult and replacement female", "Adult and replacement male", "Surplus female and male"),
  NAM = c(50, 50, 52, 55, 55, 60, 49, 50, 55),
  RUS = c(50, 50, 52, 55, 55, 60, 49, 50, 55),
  WE = c(50, 50, 52, 55, 55, 60, 49, 50, 55),
  EE = c(50, 50, 52, 55, 55, 60, 49, 50, 55),
  NENA = c(48, 48, 50, 50, 50, 55, 49, 50, 55),
  ESEA = c(50, 50, 55, 50, 50, 55, 49, 50, 55),
  OCE = c(50, 50, 52, 50, 50, 55, 49, 50, 55),
  SA = c(50, 50, 55, 50, 50, 55, 49, 50, 55),
  LAC = c(50, 50, 52, 50, 50, 55, 49, 50, 55),
  SSA = c(47, 47, 47, 47, 47, 47, 49, 50, 55)
)

dressing_percentages

# select only Adult and replacement female rows
dressing_percentages <- dressing_percentages %>%
  filter(Type == "Adult and replacement female",
         Category != "Buffaloes") |> 
  dplyr::select(-Type)
dressing_percentages





# ---------------------------------------------------------------------
# Combine with reg_wgs values
reg_wgs$Region_new

# Määritellään alueiden vastaavuudet reg_wgs ja dressing_percentages välillä
alueiden_vastaavuudet <- list(
  "Australia and Oceania" = "OCE",
  "Central America" = "LAC", # LAC kattaa laajemmin keski-Amerikan
  "Eastern Asia" = "ESEA", # ESEA kattaa Itä- ja Kaakkois-Aasian
  "Eastern Europe and Central Asia" = "EE", # EE kattaa itä-Euroopan ja voisi olettaa kattavaksi myös keski-Aasian
  "Ice" = NA, # Ei vastaavuutta, asetetaan NA
  "Indian Subcontinent" = "SA", # SA kattaa Etelä-Aasian
  "Latin America" = "LAC", # LAC kattaa Latinalaisen Amerikan
  "Middle East" = "NENA", # NENA kattaa Lähi-Idän ja Pohjois-Afrikan
  "Middle and South Africa" = "SSA", # SSA kattaa Saharan eteläpuolisen Afrikan
  "North Africa" = "NENA", # NENA kattaa Lähi-Idän ja Pohjois-Afrikan
  "North America" = "NAM", # NAM kattaa Pohjois-Amerikan
  "Southeastern Asia" = "ESEA", # ESEA kattaa Itä- ja Kaakkois-Aasian
  "Western Europe" = "WE" # WE kattaa Länsi-Euroopan
)




# Ensin määritellään maitokarjan ja nautakarjan teurastusprosentit
dairy_percentages <- c(
  "Australia and Oceania" = 50,
  "Central America" = 50, 
  "Eastern Asia" = 50,
  "Eastern Europe and Central Asia" = 50,
  "Ice" = NA, 
  "Indian Subcontinent" = 50,
  "Latin America" = 50,
  "Middle East" = 50,
  "Middle and South Africa" = 47,
  "North Africa" = 48,
  "North America" = 50,
  "Southeastern Asia" = 50,
  "Western Europe" = 50
)

beef_percentages <- c(
  "Australia and Oceania" = 55,
  "Central America" = 55, 
  "Eastern Asia" = 55,
  "Eastern Europe and Central Asia" = 55,
  "Ice" = NA, 
  "Indian Subcontinent" = 50,
  "Latin America" = 50,
  "Middle East" = 55,
  "Middle and South Africa" = 47,
  "North Africa" = 50,
  "North America" = 55,
  "Southeastern Asia" = 50,
  "Western Europe" = 55
)

# Yhdistetään teurastusprosentit reg_wgs-objektiin käyttäen `mutate`-funktiota
reg_wgs_dressing <- reg_wgs %>%
  mutate(
    dresspercDairy = dairy_percentages[Region_new] / 100,
    dresspercBeef = beef_percentages[Region_new] / 100
  )

# Tulostetaan lopullinen tietorakenne nähdäksemme uudet sarakkeet
reg_wgs_dressing |> st_drop_geometry() # näyttää ookoolta

```

# rasteroidaan reg_wgs_dressing 


```{r}
reg_wgs_dressing_vect <-  vect(as(reg_wgs_dressing, "Spatial"))


#  rasterisointi
r_dresspercDairy <- rasterize(reg_wgs_dressing_vect, template_rast_5arcmin, "dresspercDairy")
r_dresspercBeef <- rasterize(reg_wgs_dressing_vect, template_rast_5arcmin, "dresspercBeef")


plot(c(r_dresspercDairy, r_dresspercBeef)) # 5 arcmin


```


# Rasteri lihan proteiinipitoisuuksista

```{r}
faoid_and_protein_fraction  <-
  fbs_animals_wide_filled %>%
  filter(Item == "Bovine Meat") %>% # Bovine meat instead o meat
  dplyr::select(Area.Code, protein_fraction)

## create reclassification matrix
rcl_mat_countries_to_protein_fraction <-
  matrix(c(
    faoid_and_protein_fraction$Area.Code,
    faoid_and_protein_fraction$protein_fraction),
    ncol = 2)


## create raster where fao id has been replaced with protein_fraction
r_protein_fraction_Meat_from_FAOtables  <-
  cntry_raster %>%
  classify(rcl_mat_countries_to_protein_fraction,
           others =
             (fbs_animals_wide_filled %>%
                filter(Area == "World", Item == "Bovine Meat") %>%
                pull(protein_fraction)))# use global averages for missing countries

names(r_protein_fraction_Meat_from_FAOtables) <- "protein_fraction_BovMeat"

summary(r_protein_fraction_Meat_from_FAOtables) # 0.15 on average
plot(r_protein_fraction_Meat_from_FAOtables)



```


# Convert protein yields of animals to carcass weight

```{r}
# FCR perusteinen ruhopaino per pix
r_carcassWeight_from_AGB_kg_perpix <- r_protein_from_AGB_kg_perpix / r_protein_fraction_Meat_from_FAOtables


```
