==> 6_0_protein_from_crops.Rmd <==
---
title: "HYDE and MapSPAM protein"
author: "Johannes Piipponen"
date: "`r Sys.Date()`"
output: html_document
---

!!! Lisää tähän myös kcal scripti! 




Calculate protein and kcal fof grazing livestock and crops on existing production areas.
- use HYDE grazing lands

# Crops - FAO data processing
- Select crops, calculate protein fraction and kcalPerKg for each crop and country

# !!!!!!!!!  Nämä vaihdettu kun käytetään SPAM 2020
- vuosi vaihdettu 2020, myös hieman eri tiedostopolku SPAM tuotteille
- MILL nimi vaihdettu MILL


# !!!!!!!!!  Nämä on vielä vaihdettava
- SI rainfed vaihdetaan current irrigation applied SI tilalle

```{r}
# ---------------------------------------------------------- FAOSTAT download 
  ## no need to update the data if we use only yr 2010?
foodBalanceSheets_semiraw <-
  here("Data", "Input", "faostat", "foodBalanceSheets_E_All_Data_NOFLAG.csv") %>%
  read.csv() %>%
  dplyr::select(-Element.Code,-Unit , -c(Y2010:Y2019), -c(Y2021)) #
names(foodBalanceSheets_semiraw)

# ---------------------------------------------------------- crops
my_itemlist_fao <- c(  # in mapspam webpage order
  "Wheat and products", "Rice and products","Maize and products",
  "Barley and products", "Millet and products", "Sorghum and products",
  "Cereals, Other", "Potatoes and products", "Sweet potatoes",
  "Yams", "Cassava and products", "Roots, Other",  
  "Beans", "Peas", "Pulses, Other and products",
  "Soyabeans","Groundnuts", "Coconuts - Incl Copra",
  "Bananas", "Plantains", "Fruits, other", "Vegetables" # categories for SPAM names
)

my_elementlist_fao <- c(
  "Production","Food supply quantity (kg/capita/yr)",
  "Food supply (kcal/capita/day)", "Protein supply quantity (g/capita/day)"
)


  ## filter
fbs_crops <- foodBalanceSheets_semiraw  %>% 
  filter(Item %in% my_itemlist_fao,
       Element %in% my_elementlist_fao)



 ## to wide format. Select only relevant variables
fbs_crops_wide <- fbs_crops %>% 
  pivot_wider(names_from = Element, values_from = Y2020) %>% # 2020 
  rename(prod_1000mt = "Production",
         supply.KgCapYr = "Food supply quantity (kg/capita/yr)",
         supply.KcalCapD = "Food supply (kcal/capita/day)",
         prot.gCapD = "Protein supply quantity (g/capita/day)") 



  ## add column where item names corresponds to mapsmap names
  ## see Definitions and standards - Food Balances (2010-) for the classes
  ## cannot add information for all crops like this -- add MILL, COWP, PIGE, OPUL, LENT, TROF afterwards
fbs_crops_wide <- fbs_crops_wide %>% 
  mutate(spamname = case_when(
    Item == "Wheat and products" ~ "WHEA",
    Item == "Rice and products"  ~ "RICE",
    Item == "Barley and products" ~ "BARL",
    Item == "Maize and products" ~ "MAIZ",
    Item == "Millet and products"  ~ "PMIL", 
#    Item == "Millet and products"  ~ "MILL", # ei toimi koska 2x Millet and products
    Item == "Sorghum and products"  ~ "SORG",
    Item == "Cereals, Other"   ~ "OCER",
    Item == "Potatoes and products" ~ "POTA",
    Item == "Sweet potatoes"  ~ "SWPO",
    Item == "Yams"  ~ "YAMS",
    Item == "Cassava and products"   ~ "CASS",    
    Item == "Roots, Other"  ~ "ORTS",
    Item == "Beans"  ~ "BEAN",
    Item == "Pulses, Other and products"  ~ "CHIC", 
#    Item == "Pulses, Other and products" ~ "COWP", 
#    Item == "Pulses, Other and products" ~ "PIGE",     
#    Item == "Pulses, Other and products"  ~ "OPUL", 
#    Item == "Pulses, Other and products"  ~ "LENT", 
    Item == "Soyabeans"  ~ "SOYB",
    Item == "Groundnuts"   ~ "GROU",
    Item == "Coconuts - Incl Copra"  ~ "CNUT",
    Item == "Bananas"  ~ "BANA",  
    Item == "Plantains"  ~ "PLNT",
    Item == "Fruits, other"  ~ "TROF",    
#    Item == "Fruits, other"  ~ "TEMF",    
    Item == "Vegetables"  ~ "VEGE"
  ))


 # could be done with function but was faster like this..
additional_rows <- list(
  fbs_crops_wide %>% filter(Item == "Millet and products") %>% mutate(spamname = "MILL"),
  fbs_crops_wide %>% filter(Item == "Pulses, Other and products") %>% mutate(spamname = "COWP"),
  fbs_crops_wide %>% filter(Item == "Pulses, Other and products") %>% mutate(spamname = "PIGE"),
  fbs_crops_wide %>% filter(Item == "Pulses, Other and products") %>% mutate(spamname = "OPUL"),
  fbs_crops_wide %>% filter(Item == "Pulses, Other and products") %>% mutate(spamname = "LENT"),
  fbs_crops_wide %>% filter(Item == "Fruits, other") %>% mutate(spamname = "TEMF")
)

fbs_crops_wide <- fbs_crops_wide %>% bind_rows(additional_rows)
    
    
    
    
    
    
mylist_spamnames <- c(
  "WHEA", "RICE", "BARL", "MAIZ", "PMIL",
  "MILL","SORG",  "OCER", "CASS", "POTA",
  "SWPO", "YAMS", "ORTS", "BEAN", "CHIC",
  "COWP", "PIGE", "OPUL", "LENT", "SOYB",
  "GROU", "CNUT", "BANA", "PLNT", "TROF", 
  "TEMF", "VEGE") # 27 




# ------------------------- Calculate fraction of protein and kcal in a product (conversion factor) 
  ## protein and kcal
fbs_crops_wide <- fbs_crops_wide %>% 
  mutate(protein_fraction = (prot.gCapD * 365 / 1000) / supply.KgCapYr,
         kcalPerKg = (supply.KcalCapD * 365)/ supply.KgCapYr)  %>% 
  mutate(one_kg_prot_to_kcal =  (supply.KcalCapD / prot.gCapD) * 1000) |> 
  
  ## remove cases where supply = 0 (cases that yield NA values)  
  ## # also converts NaN to NA
  mutate_all(function(x) ifelse(is.nan(x) | is.infinite(x), NA, x)) 




# ---------------------------------------------------------- Fill empty rows
# Fill empty protein_fraction and kcalPerKg rows with global average values
  ## example for soy
# prot_frac_soy <- fbs_crops_wide %>% 
#   filter(spamname == "SOYB", Area == "World") %>%
#   pull(protein_fraction)
# 
# test_soy  <- fbs_crops_wide %>% 
#   filter(spamname == "SOYB") %>% 
#   mutate(protein_fraction = 
#            ifelse(is.na(protein_fraction) | protein_fraction == 0 , # if protein fraction = NA or 0,
#                                    prot_frac_soy, # fill with the global average value
#                                    protein_fraction))  # otherwise let it just be


# Create a function that fills empty values of cols protein_fraction and kcalPerKg
f_fill_missing_nutritient_values_crops <- function(cropname_mapspam) {
  
  ## pull protein fraction for spesific crop
  protein_fraction_spamcrop <- fbs_crops_wide %>% 
  filter(spamname == cropname_mapspam, Area == "World") %>%
  pull(protein_fraction)
  
  
  ## pull KcalPerKg for spesific crop
  KcalPerKg_spamcrop <- fbs_crops_wide %>% 
  filter(spamname == cropname_mapspam, Area == "World") %>%
  pull(kcalPerKg)
  
  
  ## fill empty protein fraction and kcalPerKg with global averages
  fbs_crops_wide_filled <- fbs_crops_wide %>% 
    filter(spamname == cropname_mapspam) %>%
    ## fill protein_fraction column
    mutate(protein_fraction = 
           ifelse(is.na(protein_fraction) | protein_fraction == 0 , # if protein fraction = NA or 0,
                                   protein_fraction_spamcrop, # fill with the global average value
                                   protein_fraction)) %>% # otherwise let it just be
    
    ## fill KcalPerKg column
    mutate(kcalPerKg = 
             ifelse(is.na(kcalPerKg) | kcalPerKg == 0 , 
                    KcalPerKg_spamcrop,
                    kcalPerKg)) 
  
  return(fbs_crops_wide_filled)
  }



test1 <- f_fill_missing_nutritient_values_crops("SOYB") # returns a df

# Lapply this for all the crops --> returns a list
fbs_crops_wide_filled_list <- 
  lapply(mylist_spamnames, f_fill_missing_nutritient_values_crops) 




# next we use bind_rows to make this reasonable looking dataframe
fbs_crops_wide_filled <- bind_rows(fbs_crops_wide_filled_list)

#fbs_crops_wide_filled$spamname %>% unique()


# Edit: maybe not needed as in 2020 these are separated in FAOSTAT
# Add Sudan rows to South Sudan (give new FAO id etc to South Sudan but use other values from Sudan)
# sudan <- fbs_crops_wide_filled %>% 
#   filter(Area == "Sudan") %>% 
#   dplyr::select(-Area.Code, -Area)
# 
# south_sudan <- sudan %>% 
#   mutate(Area.Code = 277,
#          Area = "South Sudan")
# 
# # Combine to the fbs_crops_wide_filled
# fbs_crops_wide_filled <- 
#   rbind(fbs_crops_wide_filled, south_sudan)
```


# Physical areas for crops 

- needed to calculate fraction of cell covered by crops
- according to the MapSPAM publication:
"PHYSICAL AREA: it is measured in hectares and represents the actual area where a crop is grown, not counting how often production was harvested from it. Physical area is calculated for each production system and crop, and the sum of all physical areas of the four production systems constitutes the total physical area for that crop. The sum of the physical areas of all crops in a pixel may not be larger than the pixel size."


```{r}
#---------------------------------- get physical areas for spam crops 
  ## (in ha originally)
f_physical_area_crops_ha_perpix <- function(cropname_mapspam) {
  list.files(path = here("Data", "Input", "mapspam", "SPAM2020", "physical_area") ,
           pattern = paste0("_", cropname_mapspam, "_A"), 
           full.names = TRUE)  %>%  
  rast()
}
f_physical_area_crops_ha_perpix("WHEA") 

# f_physical_area_crops_ha_perpix <- function(cropname_mapspam) {
#   list.files(path = here("Data", "Input", "mapspam", "physical_area") ,
#            pattern = paste0("_", cropname_mapspam, "_A"), 
#            full.names = TRUE)  %>%  ############################# 2010
#   rast()
# }


  ## get physical areas for all of the spam crops
r_physical_areas_ha_perpix <- 
  lapply(mylist_spamnames, f_physical_area_crops_ha_perpix) %>% 
  rast()# 27
names(r_physical_areas_ha_perpix) <- mylist_spamnames

  ## ex plot
#plot(r_physical_areas_ha_perpix$WHEA) 


  ## sum of these areas, read as "area of crop species in cell combined in km2"
r_physical_areas_crops_sum_ha_perpix <- sum(r_physical_areas_ha_perpix, na.rm = T) # areas per grid cell. e.g in India lot of area covered by some crops
plot(r_physical_areas_crops_sum_ha_perpix, main = "physical areas (_A_) of SPAM food crops, ha per pix)")
global(r_physical_areas_crops_sum_ha_perpix, fun = "sum", na.rm = T)/1e6 # 1043 million ha in 2020

# (in 2010 was  9.75 milj km2 total --> higher than Total global cropland area estimate of)
# https://www.statista.com/statistics/201774/projection-for-total-global-cropland-area-from-2010/
# , which estimates it was 859 million ha in 2010 ( = 8.6 milj km2) --> do not include fruits etc




# fraction of crops in 5arcmin cell
r_fraction_cl <-
  r_physical_areas_crops_sum_ha_perpix / 
  cellSize(r_physical_areas_crops_sum_ha_perpix, unit = "ha") 
names(r_fraction_cl) <- "fraction_crops"


r_fraction_cl[r_fraction_cl >1] <- 1
r_fraction_cl_0toNA <- classify(r_fraction_cl, cbind(0,NA))
r_fraction_cl_NAto0 <- classify(r_fraction_cl, cbind(NA,0))

# plot. e.g in India a large share of land area is cultivated for crops 
plot(r_fraction_cl, main = "fraction of 27 crops in 5arcmin cells") ##!!  proper plot in supplementary 



```







# Convert SPAM crop production to protein 

- according to the MapSPAM publication:
"PRODUCTION: for each production system and crop, production is calculated by multiplying area harvested by its corresponding yield. It is measured in metric tons. The total production of a crop includes the production of all production systems of that crop."


```{r}
  ## get the pdoruction data for all food crops with function
f_global_prod_crops_mt <- function(cropname_mapspam) {
  list.files(path = here("Data", "Input", "mapspam", "SPAM2020", "global_prod") ,
             pattern = paste0("P_", cropname_mapspam, "_A"),
             full.names = TRUE)  %>%
    rast()
}

f_global_prod_crops_mt("WHEA") # returns raster, unit = mt per pixel (tons per pix)

# f_global_prod_crops_mt <- function(cropname_mapspam) {
#   list.files(path = here("Data", "Input", "mapspam", "global_prod") ,
#              pattern = paste0("P_", cropname_mapspam, "_A"),
#              full.names = TRUE)  %>%
#     rast()
# }


  ## 27 global production rasters in same stack 
r_global_prod_crops_mt<- lapply(mylist_spamnames, f_global_prod_crops_mt) %>%
  rast()

# global(r_global_prod_crops_mt, fun = "sum", na.rm = T)/1e6 # global wheat yield seems to be correct, around 675 Mt
names(r_global_prod_crops_mt) <- mylist_spamnames


 ##  sum: how much prot in cell total ! Double check from methodology that it is perpixel
r_global_production_crops_sum_mt <- 
  sum(r_global_prod_crops_mt, na.rm = T)







#------------------------------------------------------- crop protein production
  ## create a function to convert crop production to protein
f_crop_protein_production_mt_perpix <- function(cropname_mapspam) {
  
  ## get protein fraction for spesific crop species
  faoid_and_prot_frac <-
    fbs_crops_wide_filled %>%
    filter(spamname == cropname_mapspam) %>%
    dplyr::select(Area.Code, protein_fraction)
  
  ## reclassification matrix
  rcl_mat_countries_to_prot_frac <-
    matrix(c(
      faoid_and_prot_frac$Area.Code,
      faoid_and_prot_frac$protein_fraction),
      ncol = 2)
  
  ## classify. Creates a raster that shows protein fraction for each country
  
    # first define values for others
  fbs_avg_proteinfract_of_a_crop <-
  fbs_crops_wide_filled %>%
  filter(spamname == cropname_mapspam, Area == "World") %>%
  pull(protein_fraction)
  
  r_fraction_prot  <- 
    cntry_raster %>% #  south sudan?
    classify(rcl_mat_countries_to_prot_frac, 
             others = fbs_avg_proteinfract_of_a_crop)
  
  
  # production raster  
  r_production_crop <-
    subset(r_global_prod_crops_mt, cropname_mapspam) 
  
  
  ## protein production raster
  r_prot <-
    r_production_crop * r_fraction_prot
  names(r_prot) <- paste0(cropname_mapspam, "_prot")
  
  return(r_prot)
}

f_crop_protein_production_mt_perpix("SOYB") 




################### lapply for all items -- 15sec

r_prot_allcrops_mt_perpix <- 
  lapply(mylist_spamnames,
         f_crop_protein_production_mt_perpix) %>% 
  rast() 


 ##  sum: how much prot in cell total 
r_prot_allcrops_sum_mt_perpix <- 
  sum(r_prot_allcrops_mt_perpix, na.rm = T)

 ## resample to get complete ext and 0 0 origin
r_prot_allcrops_sum_mt_perpix <-
  terra::resample(r_prot_allcrops_sum_mt_perpix, 
           template_rast_5arcmin) 

plot(r_prot_allcrops_sum_mt_perpix, 
     main = "27 crops: sum of protein mt") # 3rd Qu = 314




# Print total production and protein & kcal production to tibble 
tot_sums_cropspecies <- data.frame(
  crop = mylist_spamnames,
  production_Mt = global(r_global_prod_crops_mt, fun = "sum", na.rm = T)/1e6 ,
  protein_Mt = global(r_prot_allcrops_mt_perpix, fun = "sum", na.rm = T)/1e6) #,
#  kcal_MM = global(r_kcal_allcrops_MM_perpix, fun = "sum", na.rm = T)/1e6
#)
names(tot_sums_cropspecies) <- c("crop", "prod_Mt", "protein_Mt") #kcal_MM
 # check prod of rice

# Combine production statistics from FAOSTAT
tot_sums_cropspecies <- 
  left_join(tot_sums_cropspecies,
            (fbs_crops_wide_filled %>% 
                    filter(Area == "World") %>% 
                    mutate(prod_Mt_FAO = prod_1000mt/1000,
                           prot_Mt_FAO = prod_Mt_FAO*protein_fraction,
                           kcal_MM_FAO = prod_Mt_FAO * kcalPerKg/1000)) %>% 
                   dplyr::select(Item, spamname, prod_Mt_FAO, prot_Mt_FAO, kcal_MM_FAO),
  by = c("crop" = "spamname"))

tot_sums_cropspecies <- tot_sums_cropspecies %>% 
  mutate(across(where(is.numeric), round)) # !! large differences in soyabeans!


# total sums -- in 2010  production 4700 milliom tons, protein 400 Mt, energy 11 000 MM kcal 
tot_sums_cropspecies %>% 
  group_by() %>% 
  summarise(sum_production_Mt = sum(prod_Mt),
            sum_production_Mt_FAO = sum(prod_Mt_FAO), # productions (FAO & SPAM are close)
            sum_protein_Mt = sum(protein_Mt), 
            sum_protein_Mt_FAO = sum(prot_Mt_FAO)) #,
   #         sum_kcal_MM = sum(kcal_MM),
    #        sum_kcal_MM_FAO = sum(kcal_MM_FAO)) 





```




# Convert outputs to kg/ha and calculate global sums
--> note that kg/ha must include information about the fraction of cell that cl actually use! Cannot just divide by the cell area


```{r}
# crop production
r_global_production_crops_sum_kg_ha <-
  1000 * r_global_production_crops_sum_mt / 
  (cellSize(r_global_production_crops_sum_mt, unit = "ha") * r_fraction_cl_0toNA) # this must be 0 to NA as otherwise there will be div by 0
summary(r_global_production_crops_sum_kg_ha)


# crop protein production
r_prot_allcrops_sum_kg_ha <- 
  (1000 *r_prot_allcrops_sum_mt_perpix) / # convert tons to kg
  (cellSize(r_prot_allcrops_sum_mt_perpix, unit = "ha") * r_fraction_cl_0toNA)
summary(r_prot_allcrops_sum_kg_ha) # median = 198 kg of protein





# ------------------------------------------------------------------ global sums
  ## total crop prodution areas
global(r_physical_areas_crops_sum_ha_perpix, fun = "sum", na.rm = T)/1e6 # 1042 mha
  ## total production
global(r_global_production_crops_sum_mt, fun = "sum", na.rm = T)/1e6 # 6015	mmt
  ## total protein production
(global_protein_sum_mmt <-
  global(r_prot_allcrops_sum_mt_perpix, fun = "sum", na.rm = T)/1e6) # 375 mmt
#in 2010 975 million ha = 0.975 billion ha
# in 2010 total production 5374 mmt
# in 2010 total protein production 274 mmt 

# to check
# sum(tot_sums_cropspecies$protein_Mt) # 267 
# sum(tot_sums_cropspecies$prot_Mt_FAO) # 306 -- FAO global average based protein
# see our world in data   https://ourworldindata.org/food-supply



###-------------------------------------------- Calculate g protein per day per capita
  # based on FAOSTAT population in 2010 was 6869581.88 * 1000 persons
(1e12 * global_protein_sum_mmt) / # grams (mmt = 1e9 kg)
  (1000* foodBalanceSheets_semiraw %>% filter(Area == "World", Item == "Population") %>% pull(Y2020)) / # persons
  365  # days in a year
# 132 g/cap/yr, was in 2010 109 g/cap/d

    # this compared to reported protein need
((1e12 * global_protein_sum_mmt) / # grams (mmt = 1e9 kg)
  (1000* foodBalanceSheets_semiraw %>% filter(Area == "World", Item == "Population") %>% pull(Y2020)) / # persons
  365 ) /# days in a year
  (0.8*70) # protein need
# ---> in 2020 2.36 x current protein need, in 2010 was 1.95 x the current protein need

    # for who many persons this would be enough?
    # in 2020 a people would need 56g * 365d = 20440 g = 20.4kg protein in a year. Our esimate would be enough for 
(1e12 * global_protein_sum_mmt) / # total protein in grams - 1 mmt = 1e9 kg = 1e12 g
    (0.8*70*365) / # 20440 grams needed in a year
  1e6 






  # combined protein from crops and livestock = 385.1329 mmt. This is 
(1e12 * 385.1329) /
   (1000* foodBalanceSheets_semiraw %>% filter(Area == "World", Item == "Population") %>% pull(Y2020)) / # persons
  365 
# 135 g/capita/day


(1e12 * 385.1329)  / # total protein in grams
    (0.8*70*365) / # 20440 grams needed in a year
  1e6 



plot(r_physical_areas_crops_sum_ha_perpix, main = "physical areas (_A_) of SPAM food crops, ha per pix)")

plot(r_fraction_cl, main = "fraction of 27 crops in 5arcmin cells")


```


# Write rasters for intermediate use

```{r}
writeRaster(r_physical_areas_ha_perpix,
            here("Data", "Intermediate_input","r_physical_areas_ha_perpix.tif"),
            overwrite = T)


writeRaster(r_physical_areas_crops_sum_ha_perpix,
            here("Data", "Intermediate_input","r_physical_areas_crops_sum_ha_perpix.tif"),
            overwrite =T)


writeRaster(r_global_production_crops_sum_kg_ha,
            here("Data", "Intermediate_input","r_global_production_crops_sum_kg_ha.tif"),
            overwrite =T)



writeRaster(r_fraction_cl,
            here("Data", "Intermediate_input","r_fraction_cl_with_27_spamcrops.tif"),
            overwrite = T)

names(r_prot_allcrops_sum_mt_perpix) <- "crop_protein_mt_perpix"
writeRaster(r_prot_allcrops_sum_mt_perpix,
            here("Data", "Intermediate_input", "protein_production_27crops_sum_mt_perpix.tif"),
            overwrite = T)


names(r_prot_allcrops_sum_kg_ha) <- "crop_protein_kg_ha"
writeRaster(r_prot_allcrops_sum_kg_ha,
            here("Data", "Intermediate_input", "protein_production_27crops_sum_kg_ha.tif"),
            overwrite =T)



```


# How much protein produced in marginally suitable, moderately suitable and highly suitable areas?

```{r}
# Get SI data

SI_5arcmin <- #
  here("Data", "Input", "from_harddrive",
       "SI_5arcmin_overall_suitability_subset1to17_2010-2039_rcp8p5_rainfed.tif") %>%
  # here("Data", "Input", "Zabel22_SI",
  #      "SI_5arcmin_overall_suitability_subset1to17_hist1980_2009_current_irr_areas_applied.tif") %>%
  rast()

# create masks: marginally, moderately, highly
SI_marginal <- 
  terra::ifel(SI_5arcmin >= 1 & SI_5arcmin < 33, SI_5arcmin, NA)
SI_moderate <- 
  terra::ifel(SI_5arcmin >= 33 & SI_5arcmin < 75, SI_5arcmin, NA)
SI_high <- 
  terra::ifel(SI_5arcmin >= 75, SI_5arcmin, NA)


# Mask crop yields to these areas
r_prot_allcrops_sum_mt_perpix_maskedtomarginal <-
  mask(r_prot_allcrops_sum_mt_perpix, SI_marginal)
r_prot_allcrops_sum_mt_perpix_maskedtomoderate <-
  mask(r_prot_allcrops_sum_mt_perpix, SI_moderate)
r_prot_allcrops_sum_mt_perpix_maskedtohigh <-
  mask(r_prot_allcrops_sum_mt_perpix, SI_high)


# Calculate global sums for each masked dataset
global_sum_maskedtomarginal <-
  global(r_prot_allcrops_sum_mt_perpix_maskedtomarginal, "sum", na.rm = TRUE)/1e6
global_sum_maskedtomoderate <-
  global(r_prot_allcrops_sum_mt_perpix_maskedtomoderate, "sum", na.rm = TRUE)/1e6
global_sum_maskedtohigh <- 
  global(r_prot_allcrops_sum_mt_perpix_maskedtohigh, "sum", na.rm = TRUE)/1e6


# compare this to the total protein output that we get from all croplands
100 * global_sum_maskedtomarginal/global_protein_sum_mmt # 14 in 2020, was 14% in 2010
100 * global_sum_maskedtomoderate/global_protein_sum_mmt # 67 in 2020, was 66% in 2010
100 * global_sum_maskedtohigh/global_protein_sum_mmt # 16 in 2020, was 15% in 2010

global_sum_maskedtomarginal +
  global_sum_maskedtomoderate + global_sum_maskedtohigh
  

```
