==> 11_figures.Rmd <==
---
title: "Figures"
author: "Johannes Piipponen"
date: "2022-11-04"
output: html_document
---

! mieti tarviiko muihinkin karttoihin colorNA = "NA", jotta rasterisolujen seinämät eivät ole harmaita





# Color palettes for plotting

Missing fig4 palettes -- edit this anyway

```{r}
#    scico_palette_show()
pal_npp <- scico(n = 5,begin = 0.6, end = 1, palette = "corkO")
# scico(n = 6, palette = "bamako", direction = -1, begin = 0.5, end = 1)
pal_agb <- scico(n = 6, palette = "bamako", direction = -1)
# pal_protein <- scico(n = 10, palette = "hawaii", begin = 0.05,end = 0.85,
#                               direction = -1) # use lajolla instead!
pal_protein_lajolla <- scico(n = 6, palette = "lajolla", begin = 0.15, end = 1, direction = -1) # previously ended 0.85


pal_share_nuuk <- scico(n = 7, palette = "nuuk",  direction = -1) 
pal_fcr <- scico(n = 6, palette = "tokyo", direction = -1)
pal_fcr <-c("#C6DBEF", "#6BAED6", "#4292C6", "#2171B5", "#08519C", "#08306B") # maybe more clear

pal_cv <-  scico(n = 6, palette = "nuuk", direction = -1)

# test
# pal_protein_test <- 
#   c("#66c2a5", "#a6d854", "#fee08b", "#fdae61", "#f46d43", "#d53e4f")
pal_share <-  
  scico(n = 6, palette = "roma", begin = 0.15, end = 0.85, direction = -1) 
pal_lur <- scico(n =5, palette = "imola", direction = -1, begin = 0.15, end = 0.99) 




# function for plotting -- raster downsample = 0 or false
create_index_map <- function(r_index, index_label,index_main_title,
                             colorpal, breakvals,
                             breaknames = NULL,
                             color_midpoint = NULL, tocrs = NA, colorNA = NULL){
  if (!is.na(tocrs)){
    r_index <- project(r_index, tocrs, mask = TRUE)
  }
  index_map <- tm_shape(r_index,
                        raster.downsample = FALSE) + ###### or 0 instead?
    tm_raster(palette = colorpal, # try style = "fixed", 
              breaks = breakvals,
              labels = breaknames,
              title = index_label,
              midpoint = color_midpoint,
              colorNA = colorNA,
              legend.is.portrait = FALSE) + # added 9.8.22
            #  legend.reverse = TRUE) + # deleted 9.8.22
    tm_layout(main.title = index_main_title,
              main.title.position = "center",
              main.title.size = 1,
              legend.bg.color = TRUE,
              legend.outside = TRUE,
              legend.title.size = 1,
              legend.text.size = 1,
              legend.outside.size = 0.2,
              legend.outside.position = "bottom", # added 9,8
              frame = FALSE)+
    tm_shape(adm10_simple_faoadded_rob) + # was reg_rob_simple
    tm_borders(col = NA,  lwd = 0.15)  # lwd was 0.33, col was "grey30",
  
  return (index_map)
} 



# Define the function to save tmap images to spesific folder
# i = index of the layer in the list
# !! Note that nameLayers changes too!
f_save_pngs <- function(i, fig_folder) {
  
  # Create the directory if it doesn't exist
  if (!dir.exists(here("Figures", fig_folder))) {
    dir.create(here("Figures", fig_folder), recursive = TRUE)
  }
  
  # Generate the plot and save it
  p_fig <- layers[[i]] + tm_layout(legend.show = FALSE)
  tmap_save(p_fig, 
            filename = here("Figures", fig_folder, paste0('fig_', nameLayers[i], '.png')), 
            width = 110, units = 'mm', dpi = 450)
}

```

# Fig 1 is the flowchart --> made in Miro board

# Fig 2

- nrow = 3, ncol = 2
- Left column: AGB yields, AGB protein, AGB energy
- Right column: Crop yields, Crop protein, Crop energy


```{r}
# a) AGB yield
r_agb_2020_in_hyde_gl_kg_ha <-
  here("Data", "Input", "from_harddrive",
       "AGB2020_withIGPB6to10or16_in_hydeareas_kg_ha_5arcmin.tif") %>% 
  rast()

(plt_agb_hyde_in2020_kg_ha <- 
    create_index_map(r_index = r_agb_2020_in_hyde_gl_kg_ha, 
                     tocrs = "ESRI:54030",
                     index_main_title = "Median aboveground biomass (AGB) in HYDE grazing lands in 2020",
                     index_label = "AGB kg/ha (with 42% biomass removal)",
                     colorpal = pal_agb, 
                     breakvals =  c(-Inf,100,250,1000,1500,2000, Inf),  
                     breaknames = c( "< 100", "100-250","250-1000",
                                     "1000-1500", "1500-2000", "> 2000")) ) 



# b) crop yield
# Total production of 27 crops
(plt_total_production_crops <-
    create_index_map(r_index = r_global_production_crops_sum_kg_ha, 
                   tocrs = "ESRI:54030",
                   index_main_title = "Combined crop yields of SPAM food crops in 2020",
                   index_label = "[kg/ha]",
                   colorpal = pal_agb,
                   breakvals = c(0,2000,4000,6000, 8000, Inf),
                   breaknames = c("< 2000","2000-4000",
                                  "4000-6000","6000-8000",  ">8000"))) 


# c) AGB protein
(plt_animalprotein_hydeareas <- 
    create_index_map(r_index = r_protein_from_AGB_kg_ha, 
                     tocrs = "ESRI:54030",
                     index_main_title = "Estimated protein production from AGB in 2020",
                     index_label = "[kg/ha]",
                     colorpal = pal_protein_lajolla,
                     breakvals = c(0, 1, 5, 10, 25, 50, Inf),
                     breaknames = c("< 1","1-5", "5-10",
                                  "10-25","25-50",  ">50"))) 


# d) Crop protein
# sum of 27 crops
# quantile(values(r_prot_allcrops_sum_kg_ha), probs = c(seq(0,1, by = 0.1666667)), na.rm =T)
(plt_crop_protein <- 
  create_index_map(r_index = r_prot_allcrops_sum_kg_ha, 
                   tocrs = "ESRI:54030",
                   index_main_title = "Protein production of crops in 2020",
                   index_label = "[kg/ha]",
                   colorpal = pal_protein_lajolla,
                   breakvals = c(0, 10, 50, 100, 250, 500, Inf),
                   breaknames = c("< 10","10-50", "50-100",
                                  "100-250","250-500",  ">500"))) 


# e) AGB energy
(plt_animal_kcal_hydeareas <- 
    create_index_map(r_index = r_kcal_from_AGB_MM_ha, 
                     tocrs = "ESRI:54030",
                     index_main_title = "Estimated energy production from AGB in 2020",
                     index_label = "[million kcal/ha]",
                     colorpal = pal_protein_lajolla,
                     breakvals = c(0, 0.02, 0.05, 0.1, 1, 5, Inf),
                     breaknames = c("<0.02","0.02-0.05",
                                  "0.05-0.1","0.1-1", "1-5", ">5"))) 


# f) Crop energy
(plt_crop_kcal <- 
    create_index_map(r_index = r_kcal_allcrops_sum_MM_ha, 
                     tocrs = "ESRI:54030",
                     index_main_title = "Energy production of crops in 2020",
                     index_label = "[million kcal/ha]",
                     colorpal = pal_protein_lajolla,
                     breakvals = c(0, 0.2, 0.5, 1, 10, 50, Inf),
                     breaknames = c("<0.2","0.2-0.5",
                                  "0.5-1","1-10", "10-50", ">50"))) 


plts_fig2 <- tmap_arrange(
  plt_agb_hyde_in2020_kg_ha,
  plt_total_production_crops,
  plt_animalprotein_hydeareas,
  plt_crop_protein,
  plt_animal_kcal_hydeareas,
  plt_crop_kcal,
  ncol = 2
)

plts_fig2 # modify more in illustrator

tmap_save(plts_fig2, filename =
            here("Figures", "Fig 2.pdf"))

tmap_save(plts_fig2, filename =
            here("Figures", "Fig 2.png"))
```




# Fig 3

- nrow = 3, ncol = 2
- total combined protein, total combined energy
- shares of total protein, shares of total energy
- LUR protein, LUR energy

```{r}
# a) total combined protein
(plt_combined_protein <- 
  create_index_map(r_index = r_protein_combined_kg_ha, 
                   tocrs = "ESRI:54030",
                   index_main_title = "", #Total protein production from grazing livestock and crops in 2020 [kg/ha]
                   index_label = "",
                   colorpal = pal_protein_lajolla, 
                   colorNA = "NA",
                   breakvals = c(0, 10, 50, 100, 250, 500, Inf),
                   breaknames = c("<10","10-50", "50-100",
                                  "100-250","250-500",  ">500"))) 
  

# b) total combined energy
(plt_combined_energy <-
  create_index_map(r_index = r_energy_combined_MM_ha, 
                   tocrs = "ESRI:54030",
                   index_main_title = "", # Total energy production from grazing livestock and crops in 2020 [million kcal/ha]
                   index_label = "",
                   colorpal = pal_protein_lajolla,
                   breakvals = c(0, 0.2, 0.5, 1, 10, 50, Inf),
                   breaknames = c("<0.2","0.2-0.5",
                                  "0.5-1","1-10", "10-50", ">50")))
  


# c) shares of total protein
# "#2574B5" "#46ADCC" "#9CE2D5" "#D1DE99" "#BE9E3F" "#A2621C"
(plt_share_of_total_protein_livestock <- 
  create_index_map(r_index = r_share_of_total_protein_livestock, 
                   tocrs = "ESRI:54030",
                   index_main_title = "", #Grazing livestock's share of combined protein production [%]
                   index_label = "",
                   colorpal = pal_share,
                   colorNA = "NA",
                   breakvals = c(0, 10, 25, 50, 75, 90, 100),
                   breaknames = c("<10","10-25", "25-50",
                                  "50-75","75-90",  ">90")))

# d) shares of total energy
(plts_r_share_of_total_energy_livestock<-  
    create_index_map(r_index = r_share_of_total_energy_livestock, 
                     tocrs = "ESRI:54030",
                     index_main_title = "", # Grazing livestock's share of combined energy production [%]
                     index_label = "",
                     colorpal = pal_share,
                     colorNA = "NA",
                     breakvals = c(0, 10, 25, 50, 75, 90, 100),
                     breaknames = c("<10", "10-25", "25-50","50-75","75-90", "90-100"))) 


# e) LUR protein
#   "#F9FD66" "#98CB6D" "#63957A" "#3C6E90" "#274DA5"
(plt_LUR_protein <-
  create_index_map(r_index = r_LUR_protein, 
                   tocrs = "ESRI:54030",
                   index_main_title = "", # Land use ratio (LUR) from protein perspective
                   index_label = "",
                   colorpal = pal_lur,
                   colorNA = "NA",
                   breakvals = c(0, 1, 5, 10, 100,  Inf),
                   breaknames = c("<1","1-5", "5-10",
                                  "10-100",">100"))) 


# f) LUR energy
(plt_LUR_energy <-
  create_index_map(r_index = r_LUR_energy, 
                   tocrs = "ESRI:54030",
                   index_main_title = "", # Land use ratio (LUR) from energy perspective
                   index_label = "",
                   colorpal = pal_lur,
                   colorNA = "NA",
                   breakvals = c(0, 1, 5, 10, 100,  Inf),
                   breaknames = c("<1","1-5", "5-10",
                                  "10-100",">100")))



# Combine
plts_fig3 <-
  tmap_arrange(
    plt_combined_protein,
    plt_combined_energy,
    plt_share_of_total_protein_livestock,
    plts_r_share_of_total_energy_livestock,
    plt_LUR_protein,
    plt_LUR_energy,
    nrow = 3,
    ncol = 2
  )
plts_fig3

tmap_save(plts_fig3, filename = here("Figures","Fig 3.pdf"))
tmap_save(plts_fig3, filename = here("Figures","Fig 3.png"))




# ---- Link to Illustrator

# Määritä kartat ja niiden nimet
layers <- list(
  plt_combined_protein, 
  plt_combined_energy, 
  plt_share_of_total_protein_livestock, 
  plts_r_share_of_total_energy_livestock, 
  plt_LUR_protein, 
  plt_LUR_energy
)

nameLayers <- c(
  "combined_protein", 
  "combined_energy", 
  "share_of_total_protein_livestock", 
  "share_of_total_energy_livestock", 
  "LUR_protein", 
  "LUR_energy"
)



# Use the f_save_pngs function with purrr::map
map(1:length(layers), ~f_save_pngs(.x, "Fig3"))  #






```





# Fig 4

- Protein & kcal increase in scenario All to CL
- Protein & kcal increase in scenario All to GL
- Combined map of these scenarios AlltoCL & AlltoGL - including 21/27 nutrients
- Combined map, including nutrients balance and allocated outside conservation areas

```{r}
breaknames_updated <- c("< 2 (no potential)", 
                        "2-5", 
                        "5-10", 
                        "> 10") 

#scico(3, palette = "bilbao", begin = 0.2, end = 0.7, direction = -1)
pal_AlltoGL <- c("#DADADA", "#B9AE81", "#A5785C", "#8D3F46") # More GL = red

#pal_AlltoCL <- c("#e0e0e0", "#9CE2D5", "#6DCBD6", "#3B567A")
#pal_AlltoCL_AlltoGL <- c("#e0e0e0", "#9CE2D5",  "#cc6633",  "#6DCBD6", "#3B567A")

#pal_AlltoCL <- scico(4, palette = "bilbao", begin = 0.2, end = 0.95, direction = -1) #

#pal_AlltoCL <- c("#E8E8E8", "#C8B48D", "#8A6242", "#A3474B")
pal_AlltoCL <- c("#DADADA", "#C0F8F0", "#5CB6C3", "#2F3F61")
#scales::show_col(pal_AlltoCL)
pal_AlltoCL_AlltoGL <- c("#DADADA", "#C0F8F0",  "#B9AE81",  "#5CB6C3", "#2F3F61")



# All to CL - protein and kcals - no nutrients mask
(plt_protein_and_kcal_AlltoCL_recoded_fixed <-
  create_index_map(r_index = protein_and_kcal_AlltoCL_recoded_fixed,
                   tocrs = "ESRI:54030",
                   index_main_title = "",
                   index_label = "",
                   colorpal = pal_AlltoCL,
                   breakvals = c(-Inf, -554, 2, 778, Inf),
                   breaknames = breaknames_updated))



# All to GL - protein and kcals - no nutrients mask
(plt_protein_and_kcal_AlltoGL_recoded_fixed <-
  create_index_map(r_index = protein_and_kcal_AlltoGL_recoded_fixed,
                   tocrs = "ESRI:54030",
                   index_main_title = "",
                   index_label = "",
                   colorpal = pal_AlltoGL,
                   breakvals = c(-Inf, 6, 8, 12, Inf),
                   breaknames = breaknames_updated))




# Combined map of AlltoCL & AlltoGL - including 21/27 nutrients
#unique(AlltoCL_AlltoGL_combined_nutrients_considered_recode)
# -555 = no potential, 1 = moderate potential in AlltoCL scenario,
# 13 = very high potential in AlltoGL, 777 = high potential in AlltoCL,
# 10000 = very high potential in AlltoCL
(plt_combined_AlltoCL_AlltoGL_nutrients_considered_recode <-
  create_index_map(r_index = AlltoCL_AlltoGL_combined_nutrients_considered_recode,
                   tocrs = "ESRI:54030",
                   index_main_title = "",
                   index_label = "",
                   colorpal = pal_AlltoCL_AlltoGL,
                   breakvals = c(-Inf, -554, 2, 14, 778, Inf),
                   breaknames = c("< 2 (no potential)",
                                  "2-5, AlltoCL",
                                  "> 2, AlltoGL",
                                  "5-10, AlltoCL",
                                  "> 10, AlltoCL")))


# Combined map, nutrients, SIgte33
(plt_combined_AlltoCL_AlltoGL_nutrients_considered_recode_SI33 <-
  create_index_map(r_index = AlltoCL_AlltoGL_combined_nutrients_considered_recode_SI33,
                   tocrs = "ESRI:54030",
                   index_main_title = "",
                   index_label = "",
                   colorpal = pal_AlltoCL_AlltoGL,
                   breakvals = c(-Inf, -554, 2, 14, 778, Inf),
                   breaknames = c("< 2 (no potential)",
                                  "2-5, AlltoCL",
                                  "> 2, AlltoGL",
                                  "5-10, AlltoCL",
                                  "> 10, AlltoCL")))


# Further, exluding conservation areas
(plt_AlltoCL_AlltoGL_combined_nutrients_SI33_conserv_areas <-
  create_index_map(r_index = AlltoCL_AlltoGL_combined_nutrients_SI33_conserv_areas,
                   tocrs = "ESRI:54030",
                   index_main_title = "",
                   index_label = "",
                   colorpal = pal_AlltoCL_AlltoGL,
                   breakvals = c(-Inf, -554, 2, 14, 778, Inf),
                   breaknames = c("< 2 (no potential)",
                                  "2-5, AlltoCL",
                                  "> 2, AlltoGL",
                                  "5-10, AlltoCL",
                                  "> 10, AlltoCL")))


# Combine
plts_fig4 <-
  tmap_arrange(
    plt_protein_and_kcal_AlltoCL_recoded_fixed,
    plt_protein_and_kcal_AlltoGL_recoded_fixed,
    plt_combined_AlltoCL_AlltoGL_nutrients_considered_recode,
    plt_combined_AlltoCL_AlltoGL_nutrients_considered_recode_SI33,
    plt_AlltoCL_AlltoGL_combined_nutrients_SI33_conserv_areas,
    ncol = 2
  )
plts_fig4
# Save
tmap_save(plts_fig4, filename = here("Figures","Fig 4.pdf"))
tmap_save(plts_fig4, filename = here("Figures","Fig 4.png"))


# Testing! Matin koodi
# Luo kansio Fig4 Figures-kansion sisään, jos sitä ei ole olemassa
# if (!dir.exists(here("Figures", "Fig4"))) {
#   dir.create(here("Figures", "Fig4"))
# }

# Määritä kartat ja niiden nimet
layers <- list(
  plt_protein_and_kcal_AlltoCL_recoded_fixed, 
  plt_protein_and_kcal_AlltoGL_recoded_fixed, 
  plt_combined_AlltoCL_AlltoGL_nutrients_considered_recode, 
  plt_combined_AlltoCL_AlltoGL_nutrients_considered_recode_SI33,
  plt_AlltoCL_AlltoGL_combined_nutrients_SI33_conserv_areas
)

nameLayers <- c(
  'protein_and_kcal_AlltoCL', 
  'protein_and_kcal_AlltoGL', 
  'combined_nutrients', 
  'combined_nutrients_SIgte33',
  'combined_nutrients_SIgte33_conserv_areas'
)

# # Luo ja tallenna kuvat käyttäen for-silmukkaa
# for (i in 1:length(layers)) {
#   p_fig <- layers[[i]] + tm_layout(legend.show = FALSE)
#   tmap_save(p_fig, filename = here("Figures", "Fig4", paste0('fig_', nameLayers[i], '.png')), width = 110, units = 'mm', dpi = 450)
# }


map(1:length(layers), ~f_save_pngs(.x, "Fig4")) 

```



# Reality check maps (priority conservation areas)  --- AND gl, cl, mosaic

```{r}
# priority conservation areas
r_areas_where_conversion_not_realistic_combined_glmask_5arcmin <-
  rast(here("Data", "Intermediate_input",
              "areas_where_conversion_not_realistic_glmask_5arcmin.tif"))



r_areas_where_conversion_not_realistic_combined_glmask_5arcmin_NAto0 <- 
  ifel(is.na(r_areas_where_conversion_not_realistic_combined_glmask_5arcmin), 
       0,
       r_areas_where_conversion_not_realistic_combined_glmask_5arcmin) |> 
  mask(r_fraction_gl_cl_total_0toNA)


(plt_priority_conservation_areas <- 
  create_index_map(r_index = r_areas_where_conversion_not_realistic_combined_glmask_5arcmin_NAto0, 
                   tocrs = "ESRI:54030",
                   index_main_title = "Conservation priority regions",
                   index_label = "",
                   colorpal = c("#696969", "#FF9933"), 
                   breakvals =  c(-Inf, 0.1, Inf ), 
                   breaknames = c("Other areas", "Priority area"))
)

tmap_save(plt_priority_conservation_areas,
          here("Figures", "Fig priority conservation areas.pdf"))





# ------------------------------------------------------- shares of gl cl mosaic
threshold_gl_or_cl_shareover <- 80
threshold_mosaic_share_gte <- 20

# Grazing land osuus maatalousmaasta
r_gl_share_of_agrland <-  100 * r_fraction_gl_NAto0 / r_fraction_gl_cl_total_0toNA 
r_gl_share_of_agrland_over_threshold <- ifel(r_gl_share_of_agrland <= threshold_gl_or_cl_shareover, NA, r_gl_share_of_agrland)

# Croplandin osuus maatalousmaasta
r_cl_share_of_agrland <-  100 * r_fraction_cl_NAto0 / r_fraction_gl_cl_total_0toNA 
r_cl_share_of_agrland_over_threshold <- ifel(r_cl_share_of_agrland <= threshold_gl_or_cl_shareover, NA, r_cl_share_of_agrland)

# mosaic: threshold_mosaic_share_gte% tai enemmän maatalousmaasta molempia
r_gl_mask_over_threshold_mosaic <- ifel(r_gl_share_of_agrland >= threshold_mosaic_share_gte, 1, NA)
r_cl_mask_over_threshold_mosaic <- ifel(r_cl_share_of_agrland >= threshold_mosaic_share_gte, 1, NA)

# Yhdistä maskit
r_mosaic_mask <- r_gl_mask_over_threshold_mosaic * r_cl_mask_over_threshold_mosaic

# Määritä mosaic-alueet
r_gl_cl_share_mosaic <- ifel(r_mosaic_mask == 1, 1, NA)

plot(r_gl_share_of_agrland_over_threshold, main = paste("Grazing land share of agr land >", threshold_gl_or_cl_shareover, "%"))
plot(r_gl_cl_share_mosaic, main = paste("mosaic system: At least", threshold_mosaic_share_gte, "% both gl and cl"))
plot(r_cl_share_of_agrland_over_threshold, main = paste("Cropland share of agr land >", threshold_gl_or_cl_shareover, "%"))

# give new names
r_gl_share_of_agrland_over80 <- r_gl_share_of_agrland_over_threshold
r_cl_share_of_agrland_over80 <- r_cl_share_of_agrland_over_threshold



# Grazing land osuus maatalousmaasta (over 80%)
(plt_gl_share_of_agrland_over80 <- 
  create_index_map(r_index = r_gl_share_of_agrland_over80,
                   tocrs = "ESRI:54030",
                   index_main_title = "Grazing land share of agricultural land >80%",
                   index_label = "",
                   colorpal = "#A6CEE3",
                   breakvals = c(-Inf, threshold_gl_or_cl_shareover, Inf),
                   breaknames = c("", "")) )

# Cropland osuus maatalousmaasta (over 80%)
(plt_cl_share_of_agrland_over80 <- 
  create_index_map(r_index = r_cl_share_of_agrland_over80,
                   tocrs = "ESRI:54030",
                   index_main_title = "Cropland share of agricultural land >80%",
                   index_label = "",
                   colorpal = "#C0A2C7",
                   breakvals = c(-Inf, threshold_gl_or_cl_shareover, Inf),
                   breaknames = c("", "")) )

# Mosaic areas (both grazing and cropland over threshold_mosaic_share_gte%)
(plt_gl_cl_share_mosaic <- 
  create_index_map(r_index = r_gl_cl_share_mosaic,
                   tocrs = "ESRI:54030",
                   index_main_title = "Mosaic system: at least 20% both production systems",
                   index_label = "",
                   colorpal = "#1F78B4",
                   breakvals = c(-Inf, 0.5, Inf),
                   breaknames = c("", "")) )



plt_land_use_categories <- 
  tmap_arrange(plt_gl_share_of_agrland_over80, 
               plt_cl_share_of_agrland_over80, 
               plt_gl_cl_share_mosaic,
               ncol = 1
               )

tmap_save(plt_land_use_categories,
          here("Figures", "Fig land use categories.pdf"))
```




# OLD Fig 4 Line graph --> MOVE TO SUPPLEMENTARY

Plot line graph with ggplot and add LUR here

```{r}
# Get data
df_SI_positive <-
  read_excel(here("Data", "Output", "Total_sums_cropland_expansion_and_optimised_protein_production_positive.xlsx"))
           
# round
df_rounded <- round(df_SI_positive, 1) %>% 
  dplyr::select(-Protein_from_AGB_decrease_mmt)

#add median LUR value here --> not needed
# df_LUR_SI_based_NAto0_noXY_medians_positive <-
#   read.csv(here("Data", "Output", "LUR_medians_different_SI_thresholds.csv"))

df_rounded <-
  left_join(df_rounded,
            df_LUR_SI_based_NAto0_noXY_medians_positive,
            by = c("SI_class" = "SI"))

  
# Revision: change names from current to existing
df_rounded <- df_rounded %>% 
  rename("Existing grazingland area" = "Current_grazingland_area_total_Mha",
         "Existing cropland area" = "Current_cropland_area_total_Mha",
         "Existing agricultural area" = "Current_agricultural_total_area_Mha",
         "SI based unrestricted cropland" = "SI_based_unrestricted_cropland_Mha",
         "SI based restricted cropland" = "SI_based_restricted_max_cropland_Mha",
         "Potential cropland expansion" = "potential_cropland_expansion_Mha",
         "Potential cropland expansion outside cropland" = "potential_cropland_expansion_outside_cl_Mha",
         "Potential cropland expansion on grid cell containing cropland" = "potential_cropland_expansion_cl_mask_Mha",
         "Original protein from AGB" = "Protein_from_AGB_original_mmt",
         "Optimised protein from AGB" = "Protein_from_AGB_optimised_mmt",
         "Original crop protein production" = "Original_crop_protein_production_mmt",
         "Optimised crop protein" = "Optimised_crop_protein_mmt",
         "Optimised combined protein" = "Optimised_combined_protein_mmt" ,
         "Median LUR"= "LUR_median")

# Select columns
Mha_columns <- 
  c("Existing grazingland area", 
    "Existing cropland area", 
   # "Current agricultural area",
  #  "SI based unrestricted cropland", 
    "SI based restricted cropland",
    "Potential cropland expansion",
    "Potential cropland expansion outside cropland") #,
  #  "Potential cropland expansion on grid cell containing cropland")

prot_columns <- 
  c(
    #"Original protein from AGB",
    "Optimised protein from AGB",
    "Original crop protein production",
 #   "Optimised crop protein",
    "Optimised combined protein") #  "Median LUR"



# Muuta df_rounded:n pitkäksi muodoksi
df_SI_long_Mha <- df_rounded %>%
  dplyr::select(SI_class, all_of(Mha_columns)) %>%
  pivot_longer(cols = -SI_class, names_to = "Variable", values_to = "Value")

df_SI_long_protein <- df_rounded %>%
  dplyr::select(SI_class, all_of(prot_columns))  %>% 
  pivot_longer(cols = -SI_class, names_to = "Variable", values_to = "Value")

df_SI_long_LUR <- df_rounded %>%
  dplyr::select(SI_class, "Median LUR")  %>% 
  pivot_longer(cols = -SI_class, names_to = "Variable", values_to = "Value")





# Plot

# Common theme settings
library(cowplot)
common_theme <- theme_classic() +
  theme(text = element_text(size = 7),
        legend.position = "bottom",
        legend.box = "vertical",
        legend.justification = "left",
        legend.direction = "horizontal",
        legend.key.width = unit(1.5, "cm"),
        panel.spacing = unit(2, "lines"))



# Plot hectares line
plt_hectares_line <- ggplot(df_SI_long_Mha %>% filter(SI_class > 0), 
                            aes(x = SI_class, y = Value, color = Variable, 
                                linetype = ifelse(Variable == "Existing grazingland area", 
                                                  "dotted", 
                                                  ifelse(Variable %in% c("Existing cropland area", "Existing agricultural area"),
                                                         "dashed", "solid")))) +
  geom_line(linewidth = 0.7) +
  scale_x_continuous(breaks = seq(0, 100, 5), name = "SI class") +
  scale_y_continuous(name = "Area, million hectares") +
  labs(title = "",
       subtitle = "Agricultural land available for crop production based on suitability index (SI) thresholds") +
  scale_color_viridis_d() +
  scale_linetype_manual(values = c("dashed","dotted", "solid")) +
  guides(color = guide_legend(title = ""),
         linetype = guide_legend(title = "")) +  
  common_theme



# Plot protein line
plt_protein_line <- ggplot(df_SI_long_protein %>% filter(SI_class > 0), 
                           aes(x = SI_class, y = Value, color = Variable, 
                               linetype = ifelse(Variable %in% c("Original protein from AGB", "Original crop protein production"),
                                                 "dashed", "solid"))) +
  geom_line(linewidth = 0.7) +
  scale_x_continuous(breaks = seq(0, 100, 5), name = "SI class") +
  scale_y_continuous(name = "Protein, million metric tons (mmt)") +
  labs(title = "",
       subtitle = "Total protein production based on SI thresholds") +
  scale_color_viridis_d(option ="inferno", begin = 0.3, direction = -1) +
  scale_linetype_manual(values = c("dashed", "solid")) +
  guides(color = guide_legend(title = ""),
         linetype = guide_legend(title = ""))+
  common_theme


# LUR -- not needed here
# plt_LUR_line <- 
#   ggplot(df_SI_long_LUR %>% filter(SI_class > 0), 
#          aes(x = SI_class, y = Value, color = Variable)) +
#   geom_line(linewidth = 0.7) +
#   scale_x_continuous(breaks = seq(0, 100, 5), name = "SI class") +
#   scale_y_continuous(name = "Median LUR") +
#   labs(title = "",
#        subtitle = "") +
#   scale_color_scico_d(palette = "imola") +
#   theme_classic() +
#   theme(text = element_text(size = 8),
#         legend.position = "bottom",
#         panel.spacing = unit(2, "lines"))
# 
# plt_LUR_line


# Combine plots using cowplot
plt_combined <- plot_grid(plt_hectares_line, plt_protein_line, ncol = 2, align = "v")
plt_combined

# Save combined plot
ggsave(filename = here("Figures", "Figure 4 a b Linegraph.pdf"), 
       plot = plt_combined, 
       width = 12, height = 6, units = "in", 
       dpi = 300)

# ggsave(plt_hectares_line, filename = here("Figures", "Fig 4a Linegraph_potential_cl_area_hectares.pdf"))
# ggsave(plt_protein_line, filename = here("Figures", "Fig 4b Linegraph_potential_protein_production.pdf"))



```

# OLD Fig 4 (line graph polot) continuation --> MOVE TO SUPPLEMENTARY

- Plot SI based land use changes / comparisons II

- cropland expansion 33 and 75

!! Skaalan pitäisi olla > 100% saakka ja sitten uusi luokka tapauksille joissa cl expansion osuu alueille joilla ei ole aiemmin ollut croplandia. Tämän luokan nimi voisi olla "new cropland area" 

```{r}
#  ----------------------------  cropland expansion 33, 75
r_cropland_expansion_percent_lyr_SI_33_75 <- 
  subset(r_cropland_expansion_percent_positive_Inf_handled, # note -999 values that were Inf before!
         c("MaxFracOfCell_ge33", "MaxFracOfCell_ge75"))

names(r_cropland_expansion_percent_lyr_SI_33_75) <-
  c("Cropland expansion when SI threshold ≥ 33",
    "Cropland expansion when SI threshold ≥ 75")

# do this to get rid of white value that occurs when SI is too high (not sure if needed anymore..)
r_cropland_expansion_percent_lyr_SI_33_75_NAto0 <-
  classify(r_cropland_expansion_percent_lyr_SI_33_75, 
           cbind(NA,0))

  
r_cropland_expansion_percent_lyr_SI_33_75_NAto0_agrlandmask <-
  mask(r_cropland_expansion_percent_lyr_SI_33_75_NAto0,
       r_fraction_gl_cl_total_0toNA)


pal_cl_expansion <- 
  scico(10, palette = "imola", direction = -1)

# make palette out of these
pal_cl_expansion <- 
  c("#1933B2",
    "#FFFF66", "#C2E966", "#97CA6C", "#7AAD73", "#5F917A")

# pal_cl_expansion <- 
#   scico(10, palette = "nuuk", direction = -1)
# pal_cl_expansion <- 
#   c("#05598C", "#FEFEB2", "#C6C684", "#B9B88C", "#AAAC95", "#929C96")

(plt_r_cropland_expansion_percent_lyr_SI_33_75 <-
  create_index_map(r_index = r_cropland_expansion_percent_lyr_SI_33_75_NAto0_agrlandmask, 
                   tocrs = "ESRI:54030",
                   index_main_title = "Cropland expansion potential based on suitability restrictions", 
                   index_label = "[%]",
                   colorpal = pal_cl_expansion,
                     # c("#C0A2C7", "#FEFEB2", "#ADAE94", "#8D9996", "#5E7C88", "#326682", "#05598C"),
                   colorNA = "NA",
                   breakvals = c(-999 ,0, 10, 50,  100, 200, Inf),
                   breaknames = c("New cropland", "0 to 10", "10 to 50", "50 to 100", "100 to 200", "> 200")))

tmap_save(plt_r_cropland_expansion_percent_lyr_SI_33_75,
          here("Figures", "Fig 4 cd expansion maps.pdf"))

tmap_save(plt_r_cropland_expansion_percent_lyr_SI_33_75,
          here("Figures", "Fig 4 cd expansion maps.png"))






```

# OLD Fig 5 The final map --> replaced with different maps

Plotted in script 10 but moved also here

```{r}
# 25decrease to 1000% increase here:
r_protein_change_currenTO_ALLtoCL_recoded <- 
  here("Data", "Intermediate_input",
       "r_protein_change_currenTO_ALLtoCL_recoded_m25to1000.tif") |> 
  rast()

# Plot
breaknames_updated <- c("Grazing remains (29%)", 
                        "Cropland remains (14%)", 
                        "Mosaic remains (17%)", 
                        "Beneficial shifts (28%)", 
                        "Adverse shifts (12%)") #Cropland to Grazing? we dont know for sure



pal_categories <- 
  c("#A6CEE3", "#C0A2C7", "#1F78B4", "#B2DF8A", "#FFA500")#

(plt_r_protein_change_currenTO_ALLtoCL_recoded <-
create_index_map(r_index = r_protein_change_currenTO_ALLtoCL_recoded,
                 tocrs = "ESRI:54030",
                 index_main_title = "Comparing protein outputs: existing land use vs 'all suitable to cropland'",
                 index_label = "Direction of change: if protein shift [-25, 1000]%, unchanged land use",
                 colorpal = pal_categories,
                
                 breakvals = c(-Inf, -9000, -7000, -500, 10, Inf),
                 breaknames = breaknames_updated))





# ------------------------------------------------------- outside priority conservation areas
# priority conservation areas
r_areas_where_conversion_not_realistic_combined_glmask_5arcmin <-
  rast(here("Data", "Intermediate_input",
              "areas_where_conversion_not_realistic_glmask_5arcmin.tif"))


r_protein_change_currenTO_ALLtoCL_recoded_outside_conservation_areas <-
  mask(r_protein_change_currenTO_ALLtoCL_recoded,
       r_areas_where_conversion_not_realistic_combined_glmask_5arcmin,
       inverse = T)



breaknames_updated <- c("Grazing remains (11%)", 
                        "Cropland remains (29%)", 
                        "Mosaic remains (26%)", 
                        "Beneficial shifts (25%)", 
                        "Adverse shifts (9%)") #Cropland to Grazing? we dont know for sure


(plt_r_protein_change_currenTO_ALLtoCL_recoded_outside_conservation_areas <-
create_index_map(r_index = r_protein_change_currenTO_ALLtoCL_recoded_outside_conservation_areas,
                 tocrs = "ESRI:54030",
                 index_main_title = "Comparing protein outputs: existing land use vs. 'all suitable to cropland' outside priority conservation areas",
                 index_label = "Direction of change: if protein shift [-25, 1000]%, unchanged land use",
                 colorpal = pal_categories,
                 # Päivitä raja-arvot ja luokkanimet vastaamaan todellisia arvoja
                 breakvals = c(-Inf, -9000, -7000, -500, 10, Inf), # poistettu luokkia
                 breaknames = breaknames_updated))




plt_protein_change_combined <-
  tmap_arrange(plt_r_protein_change_currenTO_ALLtoCL_recoded,
               plt_r_protein_change_currenTO_ALLtoCL_recoded_outside_conservation_areas,
               ncol = 2)

plt_protein_change_combined


# Save
tmap_save(plt_protein_change_combined,
          here("Figures", "Fig XX protein shifts maps.pdf"))






```

# Extended Figures in separate script





# ------------------------------------------- OLD not needed
# Gif experiment

```{r}
#install.packages("gifski")
library(gifski)

r_cropland_expansion_percent_positive <- 
  here("Data", "Intermediate_input", "r_cropland_expansion_percent_positive.tif") %>%
  rast()

#r_cropland_expansion_percent_positive <- r_cropland_expansion_percent_positive[[10:14]]
r_cropland_expansion_percent_NAto0 <-
  classify(r_cropland_expansion_percent_positive, cbind(NA,0))

r_cropland_expansion_percent_NAto0_agrlandmask <-
  mask(r_cropland_expansion_percent_NAto0,
       r_fraction_gl_cl_total_0toNA)


pal_cl_expansion <- scico(n = 8, palette = "nuuk",  direction = -1) # try different palette?


   # r_cropland_expansion_percent_NAto0_agrlandmask 
plts_r_cropland_expansion_percent_list <-
  lapply(1:nlyr(r_cropland_expansion_percent_NAto0_agrlandmask), function(i) {
    create_index_map(r_index = r_cropland_expansion_percent_NAto0_agrlandmask[[i]],
                     tocrs = "ESRI:54030",
                     index_main_title = paste("Cropland expansion if all grazing lands with SI ≥", i, 
                                              "\ncould be used for crop production"),
                     index_label = "[Cropland expansion in %]",
                     colorpal = pal_cl_expansion,
                   breakvals = c(0, 10, 25, 50, 75, 100, 150, 200, Inf),
                   breaknames = c("0 to 10", "10 to 25", "25 to 50", "50 to 75", "75 to 100", "100 to 150", "150 to 200", "> 200"))
  })


tmap_animation(plts_r_cropland_expansion_percent_list,
                   width = 1200, height = 1000, delay = 20,dpi = 300,
                   filename = here("Figures", "Supplementary", "Cropland expansion percent animation.gif"))








# ------------------------------------------------------------------- LUR
LUR_SI_based_NAto0_rob <- project(LUR_SI_based_NAto0_positive, "ESRI:54030") # 10 min


plts_lur_list <- lapply(1:nlyr(LUR_SI_based_NAto0_rob), function(i) {
  create_index_map(r_index = LUR_SI_based_NAto0_rob[[i]],
                   index_main_title = paste("Land use ratio (LUR)",
                                            "\nwhen all agricultural land with SI ≥", i,
                                            "\ncould be used for crop production"),
                   index_label = "[LUR]",
                   colorpal = pal_lur,
                   breakvals = c(0, 1, 10, 100, Inf),
                   breaknames = c("0-1", "1-10", "10-100", ">100")) # 1h
})



tmap_animation(plts_lur_list, width = 1200, height = 1000, delay = 20,dpi = 300,
               filename = here("Figures", "Supplementary", "LUR_animation.gif"))



# Lataa magick-paketti
#install.packages("magick")
library(magick)

# Lataa molemmat gif-animaatiot kuvina
animation1 <- image_read(here("Figures", "Supplementary", "Cropland expansion percent animation.gif"))
animation2 <- image_read(here("Figures", "Supplementary", "LUR_animation.gif"))


# Yhdistä GIF-tiedostot sivuttain
combined_gif <- image_append(c(animation1, animation2))



# Tallenna yhdistetty GIF
image_write(combined_gif, here("Figures","Supplementary", "combined_animation Expansion LUR.gif"))

combined_gif2 <- image_append(c(animation1, animation2), stack = TRUE)
image_write(combined_gif2, here("Figures","Supplementary", "combined_animation Expansion LUR2.gif"))
```







# Plot SI based land use changes / comparisons II

- cropland expansion 1, 33, 75
- LUR 1, 33, 75

? need to mask outside croplands? or in croplands
--> 3 maps for each row --> 4 rows is not impossible --> 1) cl expansion 2) cl expansion on croplands 3) cl exp outside cl
4) LUR everywhere


```{r}
#  ----------------------------  cropland expansion 1, 33, 75
r_cropland_expansion_percent_lyr_SI_1_33_75 <- 
  subset(r_cropland_expansion_percent_positive, 
         c("MaxFracOfCell_ge1", "MaxFracOfCell_ge33", "MaxFracOfCell_ge75"))
names(r_cropland_expansion_percent_lyr_SI_1_33_75) <-
  c("Cropland expansion when SI threshold ≥ 1", 
    "Cropland expansion when SI threshold ≥ 33",
    "Cropland expansion when SI threshold ≥ 75")

# convert NA to 0 and then mask to gl+cl areas
r_cropland_expansion_percent_lyr_SI_1_33_75_NAto0 <-
  classify(r_cropland_expansion_percent_lyr_SI_1_33_75, cbind(NA,0))

r_cropland_expansion_percent_lyr_SI_1_33_75_NAto0_agrlandmask <-
  mask(r_cropland_expansion_percent_lyr_SI_1_33_75_NAto0,
       r_fraction_gl_cl_total_0toNA)

pal_cl_expansion <- scico(n = 8, palette = "nuuk",  direction = -1) # try different palette?


(plt_r_cropland_expansion_percent_lyr_SI_1_33_75 <-
  create_index_map(r_index = r_cropland_expansion_percent_lyr_SI_1_33_75, #r_cropland_expansion_percent_lyr_SI_1_33_75_NAto0_agrlandmask, 
                   tocrs = "ESRI:54030",
                   index_main_title = "", #Crop expansion potential based on suitability restrictions
                   index_label = "[%]",
                   colorpal = pal_cl_expansion,
                   colorNA = "NA",
                   breakvals = c(0, 10, 25, 50, 75, 100, 150, 200, Inf),
                   breaknames = c("0 to 10", "10 to 25", "25 to 50", "50 to 75", "75 to 100", "100 to 150", "150 to 200", "> 200")))






#  ----------------------------  LUR 1, 33, 75
LUR_SI_based_NAto0_lyr_SI_1_33_75 <-
  subset(LUR_SI_based_NAto0, c("LUR_SI_gte1", "LUR_SI_gte33", "LUR_SI_gte75"))


(plt_LUR_potential <- 
  create_index_map(r_index = LUR_SI_based_NAto0_lyr_SI_1_33_75, 
                   tocrs = "ESRI:54030",
                   index_main_title = "", # Potential LUR based on suitability restrictions
                   index_label = "[LUR]",
                   colorpal = pal_lur,
                   colorNA = "NA",
                   breakvals = c(0, 1, 10, 100,  Inf),
                    breaknames = c("0-1","1-10","10-100",">100"))) 





# arrange
plts_combined_cl_expansion_and_potential_LUR <-
  tmap_arrange(
    plt_r_cropland_expansion_percent_lyr_SI_1_33_75,
    plt_LUR_potential,
    ncol = 2)

plts_combined_cl_expansion_and_potential_LUR



tmap_save(plts_combined_cl_expansion_and_potential_LUR, 
          filename = here("Figures", "plts_combined_cl_expansion_and_potential_LUR.pdf"))
```



######################################### Old

# Plot comparisons part 2 - potentials --- neeed to think whether this should be presented
From suitabilities script

```{r}
# max cropland share
(plt_max_cropland_share_in_SI_class_1 <-
      create_dynamic_index_map(r_SIbased_restricted_max_cropland_fraction, 1,
                               diff_breakvals = c(-Inf, 0.10, 0.20, 0.60, 0.80, 1, Inf),
"Land area that could be used for crop production when SI threshold ≥ 1"))


(plt_max_cropland_share_in_SI_class_33 <-
      create_dynamic_index_map(r_SIbased_restricted_max_cropland_fraction, 33,
                               diff_breakvals = c(-Inf, 0.10, 0.20, 0.60, 0.80, 1, Inf),
"Land area that could be used for crop production when SI threshold ≥ 33"))


(plt_max_cropland_share_in_SI_class_75 <-
      create_dynamic_index_map(r_SIbased_restricted_max_cropland_fraction, 75,
                               diff_breakvals = c(-Inf, 0.10, 0.20, 0.60, 0.80, 1, Inf),
"Land area that could be used for crop production when SI threshold ≥ 75"))



# cropland expansion %
(plt_cropland_expansion_when_SI_class_ge_1 <- 
    create_dynamic_index_map(myraster = r_cropland_expansion_percent,
                             layer_num = 1,
                             main_title = "Cropland expansion potential when SI threshold ≥ 1"))


(plt_cropland_expansion_when_SI_class_ge_33<- 
    create_dynamic_index_map(myraster = r_cropland_expansion_percent,
                             layer_num = 33,
                             main_title = "Cropland expansion potential when SI threshold ≥ 33"))


(plt_cropland_expansion_when_SI_class_ge_75<- 
    create_dynamic_index_map(myraster = r_cropland_expansion_percent,
                             layer_num = 75,
                             main_title = "Cropland expansion potential when SI threshold ≥ 75"))




# cropland expansion % outside croplands
(plt_cropland_expansion_outside_croplands_when_SI_class_ge_1 <-
    create_dynamic_index_map(myraster = r_cropland_expansion_percent_outside_croplands,
                             layer_num = 1,
                             main_title = "Cropland expansion potential outside croplands when SI threshold ≥ 1" ))

(plt_cropland_expansion_outside_croplands_when_SI_class_ge_33 <-
    create_dynamic_index_map(myraster = r_cropland_expansion_percent_outside_croplands,
                             layer_num = 33,
                             main_title = "Cropland expansion potential outside croplands when SI threshold ≥ 33" ))


(plt_cropland_expansion_outside_croplands_when_SI_class_ge_75 <-
    create_dynamic_index_map(myraster = r_cropland_expansion_percent_outside_croplands,
                             layer_num = 75,
                             main_title = "Cropland expansion potential outside croplands when SI threshold ≥ 75" ))





# yield larger percent
plt_optimized_protein_yield_larger_when_SI_class_ge_1  <- 
  create_dynamic_index_map(r_optimized_protein_yield_larger_percent,
                           layer_num = 1,
                           main_title = "Optimized combined protein production compared to original protein production (dairy + 27 crops) when SI threshold ≥ 1")

plt_optimized_protein_yield_larger_when_SI_class_ge_33  <- 
  create_dynamic_index_map(r_optimized_protein_yield_larger_percent,
                           layer_num = 33,
                           main_title = "Optimized combined protein production compared to original protein production (dairy + 27 crops) when SI threshold ≥ 33")

plt_optimized_protein_yield_larger_when_SI_class_ge_75  <- 
  create_dynamic_index_map(r_optimized_protein_yield_larger_percent,
                           layer_num = 75,
                           main_title = "Optimized combined protein production compared to original protein production (dairy + 27 crops) when SI threshold ≥ 75")


plts_optimized <- 
tmap_arrange(
  plt_max_cropland_share_in_SI_class_1,
  plt_max_cropland_share_in_SI_class_33,
  plt_max_cropland_share_in_SI_class_75,
  plt_cropland_expansion_when_SI_class_ge_1,
  plt_cropland_expansion_when_SI_class_ge_33,
  plt_cropland_expansion_when_SI_class_ge_75,
  plt_cropland_expansion_outside_croplands_when_SI_class_ge_1,
  plt_cropland_expansion_outside_croplands_when_SI_class_ge_33,
  plt_cropland_expansion_outside_croplands_when_SI_class_ge_75,
  plt_optimized_protein_yield_larger_when_SI_class_ge_1,
  plt_optimized_protein_yield_larger_when_SI_class_ge_33,
  plt_optimized_protein_yield_larger_when_SI_class_ge_75,
  ncol = 3, nrow = 4
)



tmap_save(plts_optimized, here("Figures", "plts_optimised.pdf"))

```



```{r}

pal_AlltoCL <- c("#e0e0e0", "#9CE2D5", "#6DCBD6", "#3B567A")



pal_AlltoCL <- c("#d3d3d3", "#C0F8F0", "#5CB6C3", "#2F3F61")




# All to CL - protein and kcals - no nutrients mask
(plt_protein_and_kcal_AlltoCL_recoded_fixed <-
  create_index_map(r_index = protein_and_kcal_AlltoCL_recoded_fixed,
                   tocrs = "ESRI:54030",
                   index_main_title = "",
                   index_label = "",
                   colorpal = pal_AlltoCL,
                   breakvals = c(-Inf, -554, 2, 778, Inf),
                   breaknames = breaknames_updated))


tmap_save(plt_protein_and_kcal_AlltoCL_recoded_fixed, 
          here("del2", "test4.png"))


```
